add endPoint : getAllArticlesSection et configuration FetchType.EAGER et Lazy dans les entitées associer

+++ src/main/java/ArticleWebService/Exception/ArticleNotFoundException.java
@@ -4,7 +4,7 @@ public class ArticleNotFoundException extends RuntimeException {
 
     private static final long serialVersionUID = -2028778202108312111L;
 
-    public ArticleNotFoundException(Long id) {
+    public ArticleNotFoundException(Integer id) {
         super(String.format("L'identifiant %d  de l'article n'a pas était trouver", id));
     }
